agents/specialized/security_agent.py
# agents/specialized/security_agent.py
from typing import Dict, List, Optional
import asyncio
import logging
from datetime import datetime
from pathlib import Path

class SecurityAgent:
    def __init__(self, config: Dict = None):
        self.config = config or {}
        self.logger = logging.getLogger(__name__)
        self.vulnerability_scanner = None
        self.code_sanitizer = None
        self.compliance_checker = None

    async def analyze_security(self, project_path: Path) -> Dict:
        """Realiza análise completa de segurança"""
        try:
            results = {
                'vulnerabilities': [],
                'code_quality': [],
                'compliance': [],
                'dependencies': [],
                'timestamp': datetime.utcnow().isoformat()
            }

            # Análise de vulnerabilidades
            vuln_results = await self._scan_vulnerabilities(project_path)
            results['vulnerabilities'] = vuln_results

            # Análise de qualidade de código
            quality_results = await self._analyze_code_quality(project_path)
            results['code_quality'] = quality_results

            # Verificação de compliance
            compliance_results = await self._check_compliance(project_path)
            results['compliance'] = compliance_results

            # Análise de dependências
            dependency_results = await self._analyze_dependencies(project_path)
            results['dependencies'] = dependency_results

            return results

        except Exception as e:
            self.logger.error(f"Security analysis failed: {str(e)}")
            raise

    async def _scan_vulnerabilities(self, project_path: Path) -> List[Dict]:
        """Analisa vulnerabilidades no código"""
        try:
            results = []
            
            # Análise de padrões conhecidos
            pattern_results = await self._scan_security_patterns(project_path)
            results.extend(pattern_results)
            
            # Análise de configurações
            config_results = await self._analyze_security_configs(project_path)
            results.extend(config_results)
            
            # Análise de endpoints
            endpoint_results = await self._analyze_endpoints(project_path)
            results.extend(endpoint_results)
            
            return results
            
        except Exception as e:
            self.logger.error(f"Vulnerability scan failed: {str(e)}")
            return []

    async def _analyze_code_quality(self, project_path: Path) -> List[Dict]:
        """Analisa qualidade do código sob perspectiva de segurança"""
        try:
            results = []
            
            # Análise de complexidade
            complexity_results = await self._analyze_complexity(project_path)
            results.extend(complexity_results)
            
            # Análise de práticas seguras
            practice_results = await self._analyze_security_practices(project_path)
            results.extend(practice_results)
            
            # Análise de logs e erros
            logging_results = await self._analyze_logging(project_path)
            results.extend(logging_results)
            
            return results
            
        except Exception as e:
            self.logger.error(f"Code quality analysis failed: {str(e)}")
            return []

    async def _check_compliance(self, project_path: Path) -> List[Dict]:
        """Verifica conformidade com padrões de segurança"""
        try:
            results = []
            
            # Verificação OWASP
            owasp_results = await self._check_owasp_compliance(project_path)
            results.extend(owasp_results)
            
            # Verificação GDPR
            gdpr_results = await self._check_gdpr_compliance(project_path)
            results.extend(gdpr_results)
            
            # Verificação PCI
            pci_results = await self._check_pci_compliance(project_path)
            results.extend(pci_results)
            
            return results
            
        except Exception as e:
            self.logger.error(f"Compliance check failed: {str(e)}")
            return []

    async def _analyze_dependencies(self, project_path: Path) -> List[Dict]:
        """Analisa segurança das dependências"""
        try:
            results = []
            
            # Análise de versões
            version_results = await self._analyze_dependency_versions(project_path)
            results.extend(version_results)
            
            # Análise de licenças
            license_results = await self._analyze_licenses(project_path)
            results.extend(license_results)
            
            # Análise de CVEs
            cve_results = await self._analyze_cves(project_path)
            results.extend(cve_results)
            
            return results
            
        except Exception as e:
            self.logger.error(f"Dependency analysis failed: {str(e)}")
            return []

    # Métodos auxiliares de análise
    async def _scan_security_patterns(self, project_path: Path) -> List[Dict]:
        """Analisa padrões de segurança conhecidos"""
        patterns = [
            {'pattern': r'password\s*=', 'severity': 'high', 'description': 'Hardcoded password'},
            {'pattern': r'api_key\s*=', 'severity': 'high', 'description': 'Hardcoded API key'},
            {'pattern': r'secret\s*=', 'severity': 'high', 'description': 'Hardcoded secret'}
        ]
        return await self._scan_patterns(project_path, patterns)

    async def _analyze_security_configs(self, project_path: Path) -> List[Dict]:
        """Analisa configurações de segurança"""
        config_checks = [
            {'file': 'web.config', 'check': 'debug', 'severity': 'medium'},
            {'file': 'app.config', 'check': 'connectionStrings', 'severity': 'high'},
            {'file': 'appsettings.json', 'check': 'Authentication', 'severity': 'high'}
        ]
        return await self._check_configs(project_path, config_checks)

    async def _scan_patterns(self, project_path: Path, patterns: List[Dict]) -> List[Dict]:
        """Executa scan de padrões"""
        results = []
        for pattern in patterns:
            # Implementar lógica de scan
            pass
        return results

    async def _check_configs(self, project_path: Path, checks: List[Dict]) -> List[Dict]:
        """Executa verificações de configuração"""
        results = []
        for check in checks:
            # Implementar lógica de verificação
            pass
        return results

    async def generate_security_report(self, analysis_results: Dict) -> Dict:
        """Gera relatório de segurança"""
        try:
            report = {
                'summary': self._generate_summary(analysis_results),
                'details': self._generate_details(analysis_results),
                'recommendations': self._generate_recommendations(analysis_results),
                'timestamp': datetime.utcnow().isoformat()
            }
            return report
        except Exception as e:
            self.logger.error(f"Failed to generate security report: {str(e)}")
            raise

    def _generate_summary(self, results: Dict) -> Dict:
        """Gera sumário dos resultados"""
        return {
            'total_issues': len(results.get('vulnerabilities', [])),
            'high_severity': self._count_severity(results, 'high'),
            'medium_severity': self._count_severity(results, 'medium'),
            'low_severity': self._count_severity(results, 'low')
        }

    def _generate_details(self, results: Dict) -> List[Dict]:
        """Gera detalhes dos resultados"""
        details = []
        for category, issues in results.items():
            if isinstance(issues, list):
                for issue in issues:
                    details.append({
                        'category': category,
                        'issue': issue
                    })
        return details

    def _generate_recommendations(self, results: Dict) -> List[Dict]:
        """Gera recomendações baseadas nos resultados"""
        recommendations = []
        
        # Análise de vulnerabilidades
        if results.get('vulnerabilities'):
            recommendations.extend(self._generate_vulnerability_recommendations(
                results['vulnerabilities']
            ))
            
        # Análise de qualidade
        if results.get('code_quality'):
            recommendations.extend(self._generate_quality_recommendations(
                results['code_quality']
            ))
            
        # Análise de compliance
        if results.get('compliance'):
            recommendations.extend(self._generate_compliance_recommendations(
                results['compliance']
            ))
            
        return recommendations

    def _count_severity(self, results: Dict, severity: str) -> int:
        """Conta issues por severidade"""
        count = 0
        for category, issues in results.items():
            if isinstance(issues, list):
                count += sum(1 for i in issues if i.get('severity') == severity)
        return count